+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `VSCode` extensions for work from `tau`
 - Changed default `VSCode` theme to `Monokai`
 - Removed `C-e` binding in `VSCode` to access `ranger` in the built-in terminal
 - Added `nixify` and `flakeify` from `nix-direnv`